Calendar/scripts/Backup/requirements.txt
requests
python-dotenv
exchangelib
pycryptodome